requirements.txt
requests
beautifulsoup4
openai
gspread
google-auth